@@ -140,7 +10,92 @@
   "bugs": {
     "url": "https://github.com/akveo/ng2-smart-table/issues"
   },
-  "engines": {
-    "node": ">= 5.4.1 < 7"
+  "license": "MIT",
+  "scripts": {
+    "ng": "ng",
+    "ngh": "ngh",
+    "rimraf": "rimraf",
+    "ngc": "node_modules/.bin/ngc",
+    "start": "ng serve",
+    "start:aot": "ng serve --aot",
+    "build": "npm run clean:dist && ng build",
+    "build:prod": "npm run build -- --prod",
+    "build:prod:aot": "npm run build:prod -- --aot",
+    "build:ci": "npm run build:prod && npm run build:prod:aot && npm run build:release",
+    "build:ghpages": "npm run build:prod:aot -- --base-href \"https://akveo.github.io/ng2-smart-table/\" && ngh",
+    "build:release": "gulp build:release",
+    "test": "ng test -sr",
+    "test:coverage": "npm run clean:coverage && ng test -sr -cc",
+    "lint": "ng lint",
+    "lint:styles": "stylelint ./src/framework/**/*.scss",
+    "lint:ci": "npm run lint && npm run lint:styles",
+    "pree2e": "webdriver-manager update --standalone false --gecko false",
+    "e2e": "ng e2e",
+    "clean:dist": "npm run rimraf -- dist",
+    "changelog": "standard-changelog"
+  },
+  "dependencies": {
+    "@angular/animations": "2.4.7",
+    "@angular/common": "2.4.7",
+    "@angular/compiler": "2.4.7",
+    "@angular/core": "2.4.7",
+    "@angular/forms": "2.4.7",
+    "@angular/http": "2.4.7",
+    "@angular/platform-browser": "2.4.7",
+    "@angular/platform-browser-dynamic": "2.4.7",
+    "@angular/platform-server": "2.4.7",
+    "@angular/router": "2.4.7",
+    "core-js": "2.4.1",
+    "highlight.js": "9.10.0",
+    "lodash": "4.17.4",
+    "ng2-completer": "1.2.2",
+    "normalize.css": "5.0.0",
+    "rxjs": "5.4.0",
+    "zone.js": "0.8.10"
+  },
+  "devDependencies": {
+    "@angular/cli": "1.0.4",
+    "@angular/compiler-cli": "2.4.7",
+    "@types/chalk": "0.4.31",
+    "@types/gulp": "4.0.2",
+    "@types/highlight.js": "9.1.9",
+    "@types/jasmine": "2.5.38",
+    "@types/lodash": "4.14.58",
+    "@types/node": "6.0.63",
+    "@types/rimraf": "0.0.28",
+    "@types/run-sequence": "0.0.29",
+    "@types/rx": "2.5.33",
+    "angular-cli-ghpages": "0.5.1",
+    "chalk": "1.1.3",
+    "codelyzer": "3.0.1",
+    "gulp": "3.9.1",
+    "gulp-better-rollup": "1.1.1",
+    "gulp-clean": "0.3.2",
+    "gulp-clean-css": "3.0.4",
+    "gulp-connect": "5.0.0",
+    "gulp-htmlmin": "3.0.0",
+    "gulp-if": "2.0.2",
+    "gulp-sass": "3.1.0",
+    "gulp-sourcemaps": "2.4.1",
+    "jasmine-core": "2.5.2",
+    "jasmine-spec-reporter": "3.2.0",
+    "karma": "1.4.1",
+    "karma-chrome-launcher": "2.0.0",
+    "karma-cli": "1.0.1",
+    "karma-coverage-istanbul-reporter": "0.2.3",
+    "karma-jasmine": "1.1.0",
+    "karma-jasmine-html-reporter": "0.2.2",
+    "npm-run-all": "4.0.1",
+    "protractor": "5.1.1",
+    "resolve-bin": "0.4.0",
+    "rimraf": "2.6.1",
+    "run-sequence": "1.2.2",
+    "standard-changelog": "1.0.1",
+    "stylelint": "7.9.0",
+    "ts-node": "2.1.2",
+    "tslint": "5.2.0",
+    "tslint-eslint-rules": "4.0.0",
+    "tslint-language-service": "0.9.6",
+    "typescript": "2.3.2"
   }
 }
