@@ -9,7 +9,7 @@ export class EditCellDefault {
 
   @Output() edited = new EventEmitter<any>();
 
-  onEdited(event): boolean {
+  onEdited(event: any): boolean {
     this.edited.next(event);
     return false;
   }
@@ -19,7 +19,7 @@ export class EditCellDefault {
     return false;
   }
 
-  onClick(event): void {
+  onClick(event: any) {
     event.stopPropagation();
   }
 }
