@@ -1,24 +1,42 @@
-import { Component, Input, Output, EventEmitter, AfterViewInit } from '@angular/core';
+import { Component, Input, Output, EventEmitter, OnChanges, SimpleChanges } from '@angular/core';
 
 import { DataSource } from '../../lib/data-source/data-source';
 import { Column } from '../../lib/data-set/column';
+import { Subscription } from 'rxjs/Subscription';
 
 @Component({
   selector: 'ng2-smart-table-filter',
-  styleUrls: ['filter.scss'],
+  styleUrls: ['./filter.component.scss'],
   template: `
-    <div class="ng2-smart-filter" *ngIf="column.isFilterable">
-      <input 
-        [(ngModel)]="query"
-        (keyup)="_filter($event)"
-        [ngClass]="inputClass"
-        class="form-control"
-        type="text" 
-        placeholder="{{ column.title }}" />
+    <div class="ng2-smart-filter" *ngIf="column.isFilterable" [ngSwitch]="column.getFilterType()">
+      <select-filter *ngSwitchCase="'list'"
+                     [query]="query"
+                     [ngClass]="inputClass"
+                     [column]="column"
+                     (filter)="onFilter($event)">
+      </select-filter>
+      <checkbox-filter *ngSwitchCase="'checkbox'"
+                       [query]="query"
+                       [ngClass]="inputClass"
+                       [column]="column"
+                       (filter)="onFilter($event)">
+      </checkbox-filter>
+      <completer-filter *ngSwitchCase="'completer'"
+                        [query]="query"
+                        [ngClass]="inputClass"
+                        [column]="column"
+                        (filter)="onFilter($event)">
+      </completer-filter>
+      <input-filter *ngSwitchDefault
+                    [query]="query"
+                    [ngClass]="inputClass"
+                    [column]="column"
+                    (filter)="onFilter($event)">
+      </input-filter>
     </div>
-  `
+  `,
 })
-export class FilterComponent implements AfterViewInit {
+export class FilterComponent implements OnChanges {
 
   @Input() column: Column;
   @Input() source: DataSource;
@@ -27,39 +45,37 @@ export class FilterComponent implements AfterViewInit {
   @Output() filter = new EventEmitter<any>();
 
   query: string = '';
-  timeout: any;
-  delay: number = 300;
 
-  ngAfterViewInit(): void {
-    this.source.onChanged().subscribe((elements) => {
-      let filterConf = this.source.getFilter();
-      if (filterConf && filterConf.filters && filterConf.filters.length === 0) {
-        this.query = '';
-      }
-    });
-  }
+  protected dataChangedSub: Subscription;
 
-  _filter(event): boolean {
-    if (event.which === 13) {
-      this.addFilter();
-      // ignore tab component
-    } else if (event.which !== 9) {
-      if (this.timeout) {
-        clearTimeout(this.timeout);
+  ngOnChanges(changes: SimpleChanges) {
+    if (changes.source) {
+      if (!changes.source.firstChange) {
+        this.dataChangedSub.unsubscribe();
       }
-      this.timeout = setTimeout(() => {
-        this.addFilter();
-      }, this.delay);
+      this.dataChangedSub = this.source.onChanged().subscribe((dataChanges) => {
+        const filterConf = this.source.getFilter();
+        if (filterConf && filterConf.filters && filterConf.filters.length === 0) {
+          this.query = '';
+
+          // add a check for existing filters an set the query if one exists for this column
+          // this covers instances where the filter is set by user code while maintaining existing functionality
+        } else if (filterConf && filterConf.filters && filterConf.filters.length > 0) {
+          filterConf.filters.forEach((k: any, v: any) => {
+            if (k.field == this.column.id) {
+              this.query = k.search;
+            }
+          });
+        }
+      });
     }
-    this.filter.emit(null);
-    return false;
   }
 
-  addFilter(): void {
+  onFilter(query: string) {
     this.source.addFilter({
       field: this.column.id,
-      search: this.query,
-      filter: this.column.getFilterFunction()
+      search: query,
+      filter: this.column.getFilterFunction(),
     });
   }
 }
