@@ -1,31 +1,32 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import { Component, Input, Output, EventEmitter, OnChanges, SimpleChanges } from '@angular/core';
+import { Subscription } from 'rxjs/Subscription';
 
 import { DataSource } from '../../lib/data-source/data-source';
 
 @Component({
   selector: 'ng2-smart-table-pager',
-  styleUrls: ['pager.scss'],
+  styleUrls: ['./pager.component.scss'],
   template: `
     <nav *ngIf="shouldShow()" class="ng2-smart-pagination-nav">
       <ul class="ng2-smart-pagination pagination">
         <li class="ng2-smart-page-item page-item" [ngClass]="{disabled: getPage() == 1}">
-          <a class="ng2-smart-page-link page-link" href="#" 
+          <a class="ng2-smart-page-link page-link" href="#"
           (click)="getPage() == 1 ? false : paginate(1)" aria-label="First">
             <span aria-hidden="true">&laquo;</span>
             <span class="sr-only">First</span>
           </a>
         </li>
-        <li class="ng2-smart-page-item page-item" 
+        <li class="ng2-smart-page-item page-item"
         [ngClass]="{active: getPage() == page}" *ngFor="let page of getPages()">
-          <span class="ng2-smart-page-link page-link" 
+          <span class="ng2-smart-page-link page-link"
           *ngIf="getPage() == page">{{ page }} <span class="sr-only">(current)</span></span>
-          <a class="ng2-smart-page-link page-link" href="#" 
+          <a class="ng2-smart-page-link page-link" href="#"
           (click)="paginate(page)" *ngIf="getPage() != page">{{ page }}</a>
         </li>
-  
-        <li class="ng2-smart-page-item page-item" 
+
+        <li class="ng2-smart-page-item page-item"
         [ngClass]="{disabled: getPage() == getLast()}">
-          <a class="ng2-smart-page-link page-link" href="#" 
+          <a class="ng2-smart-page-link page-link" href="#"
           (click)="getPage() == getLast() ? false : paginate(getLast())" aria-label="Last">
             <span aria-hidden="true">&raquo;</span>
             <span class="sr-only">Last</span>
@@ -33,11 +34,10 @@ import { DataSource } from '../../lib/data-source/data-source';
         </li>
       </ul>
     </nav>
-  `
+  `,
 })
-export class PagerComponent {
+export class PagerComponent implements OnChanges {
 
-  @Input() perPage: number;
   @Input() source: DataSource;
 
   @Output() changePage = new EventEmitter<any>();
@@ -45,19 +45,27 @@ export class PagerComponent {
   protected pages: Array<any>;
   protected page: number;
   protected count: number = 0;
+  protected perPage: number;
 
-  ngOnInit(): void {
-    this.source.onChanged().subscribe((changes) => {
-      this.page = this.source.getPaging().page;
-      this.count = this.source.count();
+  protected dataChangedSub: Subscription;
 
-      if (this.isPageOutOfBounce()) {
-        this.source.setPage(--this.page);
+  ngOnChanges(changes: SimpleChanges) {
+    if (changes.source) {
+      if (!changes.source.firstChange) {
+        this.dataChangedSub.unsubscribe();
       }
+      this.dataChangedSub = this.source.onChanged().subscribe((dataChanges) => {
+        this.page = this.source.getPaging().page;
+        this.perPage = this.source.getPaging().perPage;
+        this.count = this.source.count();
+        if (this.isPageOutOfBounce()) {
+          this.source.setPage(--this.page);
+        }
 
-      this.processPageChange(changes);
-      this.initPages();
-    });
+        this.processPageChange(dataChanges);
+        this.initPages();
+      });
+    }
   }
 
   /**
@@ -66,7 +74,7 @@ export class PagerComponent {
    * if a new element was added to the beginning of the table - then to the first page
    * @param changes
    */
-  processPageChange(changes): void {
+  processPageChange(changes: any) {
     if (changes['action'] === 'prepend') {
       this.source.setPage(1);
     }
@@ -103,7 +111,7 @@ export class PagerComponent {
   }
 
   initPages() {
-    let pagesCount = this.getLast();
+    const pagesCount = this.getLast();
     let showPagesCount = 4;
     showPagesCount = pagesCount < showPagesCount ? pagesCount : showPagesCount;
     this.pages = [];
@@ -116,7 +124,7 @@ export class PagerComponent {
       let lastOne = middleOne + Math.floor(showPagesCount / 2);
       lastOne = lastOne >= pagesCount ? pagesCount : lastOne;
 
-      let firstOne = lastOne - showPagesCount + 1;
+      const firstOne = lastOne - showPagesCount + 1;
 
       for (let i = firstOne; i <= lastOne; i++) {
         this.pages.push(i);
