@@ -1,28 +1,35 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import {Component, Input, Output, EventEmitter, OnChanges, ChangeDetectionStrategy } from '@angular/core';
 
 import { Grid } from '../../../lib/grid';
 import { Row } from '../../../lib/data-set/row';
+import { DataSource } from '../../../lib/data-source/data-source';
 
 @Component({
   selector: 'ng2-st-tbody-edit-delete',
+  changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
-    <a href="#" *ngIf="grid.getSetting('actions.edit')" class="ng2-smart-action ng2-smart-action-edit-edit"
-        [innerHTML]="grid.getSetting('edit.editButtonContent')" (click)="onEdit($event)"></a>
-    <a href="#" *ngIf="grid.getSetting('actions.delete')" class="ng2-smart-action ng2-smart-action-delete-delete"
-        [innerHTML]="grid.getSetting('delete.deleteButtonContent')" (click)="onDelete($event)"></a>
-  `
+    <a href="#" *ngIf="isActionEdit" class="ng2-smart-action ng2-smart-action-edit-edit"
+        [innerHTML]="editRowButtonContent" (click)="onEdit($event)"></a>
+    <a href="#" *ngIf="isActionDelete" class="ng2-smart-action ng2-smart-action-delete-delete"
+        [innerHTML]="deleteRowButtonContent" (click)="onDelete($event)"></a>
+  `,
 })
-export class TbodyEditDeleteComponent {
+export class TbodyEditDeleteComponent implements OnChanges {
 
   @Input() grid: Grid;
   @Input() row: Row;
-  @Input() source: any;
+  @Input() source: DataSource;
   @Input() deleteConfirm: EventEmitter<any>;
   @Input() editConfirm: EventEmitter<any>;
 
   @Output() edit = new EventEmitter<any>();
   @Output() delete = new EventEmitter<any>();
   @Output() editRowSelect = new EventEmitter<any>();
+
+  isActionEdit: boolean;
+  isActionDelete: boolean;
+  editRowButtonContent: string;
+  deleteRowButtonContent: string;
 
   onEdit(event: any) {
     event.preventDefault();
@@ -33,7 +40,7 @@ export class TbodyEditDeleteComponent {
     if (this.grid.getSetting('mode') === 'external') {
       this.edit.emit({
         data: this.row.getData(),
-        source: this.source
+        source: this.source,
       });
     } else {
       this.grid.edit(this.row);
@@ -47,10 +54,17 @@ export class TbodyEditDeleteComponent {
     if (this.grid.getSetting('mode') === 'external') {
       this.delete.emit({
         data: this.row.getData(),
-        source: this.source
+        source: this.source,
       });
     } else {
       this.grid.delete(this.row, this.deleteConfirm);
     }
+  }
+
+  ngOnChanges(){
+    this.isActionEdit = this.grid.getSetting('actions.edit');
+    this.isActionDelete = this.grid.getSetting('actions.delete');
+    this.editRowButtonContent = this.grid.getSetting('edit.editButtonContent');
+    this.deleteRowButtonContent = this.grid.getSetting('delete.deleteButtonContent');
   }
 }
