@@ -1,4 +1,4 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import {Component, Input, Output, EventEmitter, OnChanges } from '@angular/core';
 
 import { Grid } from '../../../lib/grid';
 
@@ -6,13 +6,23 @@ import { Grid } from '../../../lib/grid';
   selector: 'ng2-st-actions',
   template: `
     <a href="#" class="ng2-smart-action ng2-smart-action-add-create"
-        [innerHTML]="grid.getSetting('add.createButtonContent')" (click)="$event.preventDefault();create.emit($event)"></a>
+        [innerHTML]="createButtonContent"
+        (click)="$event.preventDefault();create.emit($event)"></a>
     <a href="#" class="ng2-smart-action ng2-smart-action-add-cancel"
-        [innerHTML]="grid.getSetting('add.cancelButtonContent')" (click)="$event.preventDefault();grid.createFormShown = false;"></a>
-  `
+        [innerHTML]="cancelButtonContent"
+        (click)="$event.preventDefault();grid.createFormShown = false;"></a>
+  `,
 })
-export class ActionsComponent {
+export class ActionsComponent implements OnChanges {
 
   @Input() grid: Grid;
   @Output() create = new EventEmitter<any>();
+
+  createButtonContent: string;
+  cancelButtonContent: string;
+
+  ngOnChanges() {
+    this.createButtonContent = this.grid.getSetting('add.createButtonContent');
+    this.cancelButtonContent = this.grid.getSetting('add.cancelButtonContent');
+  }
 }
