@@ -1,19 +1,23 @@
-import { Component, Input, Output, EventEmitter, AfterViewInit, ElementRef } from '@angular/core';
+import { Component, Input, Output, EventEmitter, AfterViewInit, ElementRef, OnChanges } from '@angular/core';
 
 import { Grid } from '../../../lib/grid';
+import { DataSource } from '../../../lib/data-source/data-source';
 
 @Component({
   selector: '[ng2-st-add-button]',
   template: `
-    <a *ngIf="grid.getSetting('actions.add')" href="#" class="ng2-smart-action ng2-smart-action-add-add"
-        [innerHTML]="grid.getSetting('add.addButtonContent')" (click)="onAdd($event)"></a>
-  `
+    <a *ngIf="isActionAdd" href="#" class="ng2-smart-action ng2-smart-action-add-add"
+        [innerHTML]="addNewButtonContent" (click)="onAdd($event)"></a>
+  `,
 })
-export class AddButtonComponent implements AfterViewInit {
+export class AddButtonComponent implements AfterViewInit, OnChanges {
 
   @Input() grid: Grid;
-  @Input() source: any;
+  @Input() source: DataSource;
   @Output() create = new EventEmitter<any>();
+
+  isActionAdd: boolean;
+  addNewButtonContent: string;
 
   constructor(private ref: ElementRef) {
   }
@@ -22,14 +26,20 @@ export class AddButtonComponent implements AfterViewInit {
     this.ref.nativeElement.classList.add('ng2-smart-actions-title', 'ng2-smart-actions-title-add');
   }
 
+  ngOnChanges() {
+    this.isActionAdd = this.grid.getSetting('actions.add');
+    this.addNewButtonContent = this.grid.getSetting('add.addButtonContent');
+  }
+
   onAdd(event: any) {
     event.preventDefault();
     event.stopPropagation();
     if (this.grid.getSetting('mode') === 'external') {
       this.create.emit({
-        source: this.source
+        source: this.source,
       });
-    } else
+    } else {
       this.grid.createFormShown = true;
+    }
   }
 }
