@@ -1,6 +1,7 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 
 import { Column } from '../../../lib/data-set/column';
+import { DataSource } from '../../../lib/data-source/data-source';
 
 @Component({
   selector: 'ng2-st-column-title',
@@ -8,11 +9,13 @@ import { Column } from '../../../lib/data-set/column';
     <div class="ng2-smart-title">
       <ng2-smart-table-title [source]="source" [column]="column" (sort)="sort.emit($event)"></ng2-smart-table-title>
     </div>
-  `
+  `,
 })
 export class ColumnTitleComponent {
 
   @Input() column: Column;
-  @Input() source: any;
+  @Input() source: DataSource;
+
   @Output() sort = new EventEmitter<any>();
+
 }
