@@ -1,33 +1,48 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import {Component, Input, Output, EventEmitter, OnChanges} from '@angular/core';
 
 import { Grid } from '../../../lib/grid';
+import { DataSource } from '../../../lib/data-source/data-source';
+import { Column } from "../../../lib/data-set/column";
 
 @Component({
   selector: '[ng2-st-thead-filters-row]',
   template: `
-    <th *ngIf="grid.isMultiSelectVisible()"></th>
-    <th ng2-st-add-button *ngIf="grid.showActionColumn('left')"
+    <th *ngIf="isMultiSelectVisible"></th>
+    <th ng2-st-add-button *ngIf="showActionColumnLeft"
                           [grid]="grid"
                           (create)="create.emit($event)">
     </th>
     <th *ngFor="let column of grid.getColumns()" class="ng2-smart-th {{ column.id }}">
       <ng2-smart-table-filter [source]="source"
                               [column]="column"
-                              [inputClass]="grid.getSetting('filter.inputClass')"
+                              [inputClass]="filterInputClass"
                               (filter)="filter.emit($event)">
       </ng2-smart-table-filter>
     </th>
-    <th ng2-st-add-button *ngIf="grid.showActionColumn('right')"
+    <th ng2-st-add-button *ngIf="showActionColumnRight"
                           [grid]="grid"
-                          [source]="source"                          
+                          [source]="source"
                           (create)="create.emit($event)">
     </th>
-    `
+  `,
 })
-export class TheadFitlersRowComponent {
+export class TheadFitlersRowComponent implements OnChanges {
 
   @Input() grid: Grid;
-  @Input() source: any;
+  @Input() source: DataSource;
+
   @Output() create = new EventEmitter<any>();
   @Output() filter = new EventEmitter<any>();
+
+  isMultiSelectVisible: boolean;
+  showActionColumnLeft: boolean;
+  showActionColumnRight: boolean;
+  filterInputClass: string;
+
+  ngOnChanges() {
+    this.isMultiSelectVisible = this.grid.isMultiSelectVisible();
+    this.showActionColumnLeft = this.grid.showActionColumn('left');
+    this.showActionColumnRight = this.grid.showActionColumn('right');
+    this.filterInputClass = this.grid.getSetting('filter.inputClass');
+  }
 }
