@@ -1,4 +1,4 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import { Component, Input, Output, EventEmitter, OnChanges } from '@angular/core';
 
 import { Grid } from '../../../lib/grid';
 import { Row } from '../../../lib/data-set/row';
@@ -6,8 +6,8 @@ import { Row } from '../../../lib/data-set/row';
 @Component({
   selector: '[ng2-st-thead-form-row]',
   template: `
-      <td *ngIf="grid.isMultiSelectVisible()"></td>
-      <td class="ng2-smart-actions">
+      <td *ngIf=""></td>
+      <td  *ngIf="showActionColumnLeft"  class="ng2-smart-actions">
         <ng2-st-actions [grid]="grid" (create)="onCreate($event)"></ng2-st-actions>
       </td>
       <td *ngFor="let cell of grid.getNewRow().getCells()">
@@ -15,23 +15,39 @@ import { Row } from '../../../lib/data-set/row';
                               [grid]="grid"
                               [isNew]="true"
                               [createConfirm]="createConfirm"
-                              [inputClass]="grid.getSetting('add.inputClass')"
+                              [inputClass]="addInputClass"
                               [isInEditing]="grid.getNewRow().isInEditing"
                               (edited)="onCreate($event)">
         </ng2-smart-table-cell>
       </td>
-    `
+      <td  *ngIf="showActionColumnRight"  class="ng2-smart-actions">
+        <ng2-st-actions [grid]="grid" (create)="onCreate($event)"></ng2-st-actions>
+      </td>
+  `,
 })
-export class TheadFormRowComponent {
+export class TheadFormRowComponent implements OnChanges {
 
   @Input() grid: Grid;
   @Input() row: Row;
   @Input() createConfirm: EventEmitter<any>;
+
   @Output() create = new EventEmitter<any>();
+
+  isMultiSelectVisible: boolean;
+  showActionColumnLeft: boolean;
+  showActionColumnRight: boolean;
+  addInputClass: string;
 
   onCreate(event: any) {
     event.stopPropagation();
 
     this.grid.create(this.grid.getNewRow(), this.createConfirm);
+  }
+
+  ngOnChanges(){
+    this.isMultiSelectVisible = this.grid.isMultiSelectVisible();
+    this.showActionColumnLeft = this.grid.showActionColumn('left');
+    this.showActionColumnRight = this.grid.showActionColumn('right');
+    this.addInputClass = this.grid.getSetting('add.inputClass');
   }
 }
