@@ -1,29 +1,44 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import {Component, Input, Output, EventEmitter, OnChanges} from '@angular/core';
 
 import { Grid } from '../../../lib/grid';
+import { DataSource } from '../../../lib/data-source/data-source';
+import { Column } from "../../../lib/data-set/column";
 
 @Component({
   selector: '[ng2-st-thead-titles-row]',
   template: `
-    <th ng2-st-checkbox-select-all *ngIf="grid.isMultiSelectVisible()"
+    <th ng2-st-checkbox-select-all *ngIf="isMultiSelectVisible"
                                    [grid]="grid"
                                    [source]="source"
                                    [isAllSelected]="isAllSelected"
                                    (click)="selectAllRows.emit($event)">
     </th>
-    <th ng2-st-actions-title *ngIf="grid.showActionColumn('left')" [grid]="grid"></th>
-    <th *ngFor="let column of grid.getColumns()" class="ng2-smart-th {{ column.id }}" [ngClass]="column.class">
+    <th ng2-st-actions-title *ngIf="showActionColumnLeft" [grid]="grid"></th>
+    <th *ngFor="let column of grid.getColumns()" class="ng2-smart-th {{ column.id }}" [ngClass]="column.class"
+      [style.width]="column.width" >
       <ng2-st-column-title [source]="source" [column]="column" (sort)="sort.emit($event)"></ng2-st-column-title>
     </th>
-    <th ng2-st-actions-title *ngIf="grid.showActionColumn('right')" [grid]="grid"></th>
-    `
+    <th ng2-st-actions-title *ngIf="showActionColumnRight" [grid]="grid"></th>
+  `,
 })
-export class TheadTitlesRowComponent {
+export class TheadTitlesRowComponent implements OnChanges {
 
   @Input() grid: Grid;
   @Input() isAllSelected: boolean;
-  @Input() source: any;
+  @Input() source: DataSource;
 
   @Output() sort = new EventEmitter<any>();
   @Output() selectAllRows = new EventEmitter<any>();
+
+  isMultiSelectVisible: boolean;
+  showActionColumnLeft: boolean;
+  showActionColumnRight: boolean;
+
+
+  ngOnChanges() {
+    this.isMultiSelectVisible = this.grid.isMultiSelectVisible();
+    this.showActionColumnLeft = this.grid.showActionColumn('left');
+    this.showActionColumnRight = this.grid.showActionColumn('right');
+  }
+
 }
