@@ -1,15 +1,16 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import {Component, Input, Output, EventEmitter, OnChanges} from '@angular/core';
 
 import { Grid } from '../../lib/grid';
+import { DataSource } from '../../lib/data-source/data-source';
 
 @Component({
     selector: '[ng2-st-thead]',
-    templateUrl: './thead.component.html'
+    templateUrl: './thead.component.html',
 })
-export class Ng2SmartTableTheadComponent {
+export class Ng2SmartTableTheadComponent implements OnChanges {
 
     @Input() grid: Grid;
-    @Input() source: any;
+    @Input() source: DataSource;
     @Input() isAllSelected: boolean;
     @Input() createConfirm: EventEmitter<any>;
 
@@ -17,4 +18,12 @@ export class Ng2SmartTableTheadComponent {
     @Output() selectAllRows = new EventEmitter<any>();
     @Output() create = new EventEmitter<any>();
     @Output() filter = new EventEmitter<any>();
+
+    isHideHeader: boolean;
+    isHideSubHeader: boolean;
+
+  ngOnChanges() {
+      this.isHideHeader = this.grid.getSetting('hideHeader');
+      this.isHideSubHeader = this.grid.getSetting('hideSubHeader');
+    }
 }
