@@ -2,60 +2,77 @@ import { DataSet } from './data-set';
 
 export class Column {
 
-  public title: string = '';
-  public type: string = '';
-  public class: string = '';
-  public isSortable: boolean = false;
-  public isEditable: boolean = true;
-  public isFilterable: boolean = false;
-  public sortDirection: string = '';
-  public defaultSortDirection: string = '';
-  public editor: { type: string, config: any, component: any } = { type: '', config: {}, component: null };
+  title: string = '';
+  type: string = '';
+  class: string = '';
+  width: string = '';
+  isSortable: boolean = false;
+  isEditable: boolean = true;
+  isAddable: boolean = true;
+  isFilterable: boolean = false;
+  sortDirection: string = '';
+  defaultSortDirection: string = '';
+  editor: { type: string, config: any, component: any } = { type: '', config: {}, component: null };
+  filter: { type: string, config: any } = { type: '', config: {} };
+  renderComponent: any = null;
   compareFunction: Function;
   valuePrepareFunction: Function;
   filterFunction: Function;
-  cellRenderFunction: Function;
+  onComponentInitFunction: Function;
 
   constructor(public id: string, protected settings: any, protected dataSet: DataSet) {
     this.process();
   }
 
-  public getCompareFunction(): Function {
+  getOnComponentInitFunction(): Function {
+    return this.onComponentInitFunction;
+  }
+
+  getCompareFunction(): Function {
     return this.compareFunction;
   }
 
-  public getValuePrepareFunction(): Function {
+  getValuePrepareFunction(): Function {
     return this.valuePrepareFunction;
   }
 
-  public getFilterFunction(): Function {
+  getFilterFunction(): Function {
     return this.filterFunction;
   }
 
-  public getCellRenderFunction(): Function {
-    return this.cellRenderFunction;
+  getConfig(): any {
+    return this.editor && this.editor.config;
   }
 
-  public getConfig(): any {
-    return this.editor.config;
+  getFilterType(): any {
+    return this.filter && this.filter.type;
   }
 
-  protected process(): void {
+  getFilterConfig(): any {
+    return this.filter && this.filter.config;
+  }
+
+  protected process() {
     this.title = this.settings['title'];
     this.class = this.settings['class'];
+    this.width = this.settings['width'];
     this.type = this.prepareType();
     this.editor = this.settings['editor'];
+    this.filter = this.settings['filter'];
+    this.renderComponent = this.settings['renderComponent'];
 
     this.isFilterable = typeof this.settings['filter'] === 'undefined' ? true : !!this.settings['filter'];
-    this.defaultSortDirection = ['asc', 'desc'].indexOf(this.settings['sortDirection']) !== -1 ? this.settings['sortDirection'] : '';
+    this.defaultSortDirection = ['asc', 'desc']
+      .indexOf(this.settings['sortDirection']) !== -1 ? this.settings['sortDirection'] : '';
     this.isSortable = typeof this.settings['sort'] === 'undefined' ? true : !!this.settings['sort'];
     this.isEditable = typeof this.settings['editable'] === 'undefined' ? true : !!this.settings['editable'];
+    this.isAddable=typeof this.settings['addable'] === 'undefined' ? true : !!this.settings['addable'];
     this.sortDirection = this.prepareSortDirection();
 
     this.compareFunction = this.settings['compareFunction'];
     this.valuePrepareFunction = this.settings['valuePrepareFunction'];
     this.filterFunction = this.settings['filterFunction'];
-    this.cellRenderFunction = this.settings['cellRenderFunction'];
+    this.onComponentInitFunction = this.settings['onComponentInitFunction'];
   }
 
   prepareType(): string {
