@@ -3,7 +3,7 @@ import { Column } from './column';
 
 export class DataSet {
 
-  public newRow: Row;
+  newRow: Row;
 
   protected data: Array<any> = [];
   protected columns: Array<Column> = [];
@@ -18,7 +18,7 @@ export class DataSet {
     this.createNewRow();
   }
 
-  setData(data: Array<any>): void {
+  setData(data: Array<any>) {
     this.data = data;
     this.createRows();
   }
@@ -31,20 +31,29 @@ export class DataSet {
     return this.rows;
   }
 
-  findRowByData(data): Row {
+  getFirstRow(): Row {
+    return this.rows[0];
+  }
+
+  getLastRow(): Row {
+    return this.rows[this.rows.length - 1];
+  }
+
+  findRowByData(data: any): Row {
     return this.rows.find((row: Row) => row.getData() === data);
   }
 
-  deselectAll(): void {
+  deselectAll() {
     this.rows.forEach((row) => {
       row.isSelected = false;
     });
   }
 
   selectRow(row: Row): Row {
+    const previousIsSelected = row.isSelected;
     this.deselectAll();
 
-    row.isSelected = true;
+    row.isSelected = !previousIsSelected;
     this.selectedRow = row;
 
     return this.selectedRow;
@@ -82,11 +91,11 @@ export class DataSet {
     }
   }
 
-  willSelectFirstRow(): void {
+  willSelectFirstRow() {
     this.willSelect = 'first';
   }
 
-  willSelectLastRow(): void {
+  willSelectLastRow() {
     this.willSelect = 'last';
   }
 
@@ -109,8 +118,8 @@ export class DataSet {
     return this.selectedRow;
   }
 
-  createNewRow(): void {
-    this.newRow = new Row(0, {}, this);
+  createNewRow() {
+    this.newRow = new Row(-1, {}, this);
     this.newRow.isInEditing = true;
   }
 
@@ -119,8 +128,8 @@ export class DataSet {
    * @param settings
    * @private
    */
-  createColumns(settings) {
-    for (let id in settings) {
+  createColumns(settings: any) {
+    for (const id in settings) {
       if (settings.hasOwnProperty(id)) {
         this.columns.push(new Column(id, settings[id], this));
       }
