@@ -1,6 +1,6 @@
 export class LocalSorter {
 
-  protected static COMPARE = (direction, a, b) => {
+  protected static COMPARE = (direction: any, a: any, b: any) => {
     if (a < b) {
       return -1 * direction;
     }
@@ -8,12 +8,12 @@ export class LocalSorter {
       return direction;
     }
     return 0;
-  };
+  }
 
   static sort(data: Array<any>, field: string, direction: string, customCompare?: Function): Array<any> {
 
-    let dir: number = (direction === 'asc') ? 1 : -1;
-    let compare: Function = customCompare ? customCompare : this.COMPARE;
+    const dir: number = (direction === 'asc') ? 1 : -1;
+    const compare: Function = customCompare ? customCompare : this.COMPARE;
 
     return data.sort((a, b) => {
       return compare.call(null, dir, a[field], b[field]);
