@@ -1,10 +1,13 @@
 import { Http } from '@angular/http';
-import { LocalDataSource } from '../local/local.data-source';
 import { RequestOptionsArgs } from '@angular/http/src/interfaces';
-import { URLSearchParams } from '@angular/http/src/url_search_params';
+import { URLSearchParams } from '@angular/http';
 import { Observable } from 'rxjs';
+
+import { LocalDataSource } from '../local/local.data-source';
 import { ServerSourceConf } from './server-source.conf';
 import { getDeepFromObject } from '../../helpers';
+
+import 'rxjs/add/operator/toPromise';
 
 export class ServerDataSource extends LocalDataSource {
 
@@ -12,7 +15,7 @@ export class ServerDataSource extends LocalDataSource {
 
   protected lastRequestCount: number = 0;
 
-  constructor(protected http: Http, conf: ServerSourceConf|{} = {}) {
+  constructor(protected http: Http, conf: ServerSourceConf | {} = {}) {
     super();
 
     this.conf = new ServerSourceConf(conf);
@@ -40,15 +43,16 @@ export class ServerDataSource extends LocalDataSource {
    * @param res
    * @returns {any}
    */
-  protected extractDataFromResponse(res): Array<any> {
+  protected extractDataFromResponse(res: any): Array<any> {
     const rawData = res.json();
-    let data = !!this.conf.dataKey ? getDeepFromObject(rawData, this.conf.dataKey, []) : rawData;
+    const data = !!this.conf.dataKey ? getDeepFromObject(rawData, this.conf.dataKey, []) : rawData;
 
     if (data instanceof Array) {
       return data;
     }
 
-    throw new Error(`Data must be an array. Please check that data extracted from the server response by the key '${this.conf.dataKey}' exists and is array.`);
+    throw new Error(`Data must be an array.
+    Please check that data extracted from the server response by the key '${this.conf.dataKey}' exists and is array.`);
   }
 
   /**
@@ -57,12 +61,12 @@ export class ServerDataSource extends LocalDataSource {
    * @param res
    * @returns {any}
    */
-  protected extractTotalFromResponse(res): number {
+  protected extractTotalFromResponse(res: any): number {
     if (res.headers.has(this.conf.totalKey)) {
       return +res.headers.get(this.conf.totalKey);
     } else {
       const rawData = res.json();
-      return getDeepFromObject(rawData, this.conf.totalKey, 0)
+      return getDeepFromObject(rawData, this.conf.totalKey, 0);
     }
   }
 
@@ -72,7 +76,7 @@ export class ServerDataSource extends LocalDataSource {
 
   protected createRequestOptions(): RequestOptionsArgs {
     let requestOptions: RequestOptionsArgs = {};
-    requestOptions.search = new URLSearchParams();
+    requestOptions.params = new URLSearchParams();
 
     requestOptions = this.addSortRequestOptions(requestOptions);
     requestOptions = this.addFilterRequestOptions(requestOptions);
@@ -80,7 +84,7 @@ export class ServerDataSource extends LocalDataSource {
   }
 
   protected addSortRequestOptions(requestOptions: RequestOptionsArgs): RequestOptionsArgs {
-    let searchParams: URLSearchParams = <URLSearchParams> requestOptions.search;
+    const searchParams: URLSearchParams = <URLSearchParams>requestOptions.params;
 
     if (this.sortConf) {
       this.sortConf.forEach((fieldConf) => {
@@ -93,10 +97,10 @@ export class ServerDataSource extends LocalDataSource {
   }
 
   protected addFilterRequestOptions(requestOptions: RequestOptionsArgs): RequestOptionsArgs {
-    let searchParams: URLSearchParams = <URLSearchParams> requestOptions.search;
+    const searchParams: URLSearchParams = <URLSearchParams>requestOptions.params;
 
     if (this.filterConf.filters) {
-      this.filterConf.filters.forEach((fieldConf) => {
+      this.filterConf.filters.forEach((fieldConf: any) => {
         if (fieldConf['search']) {
           searchParams.set(this.conf.filterFieldKey.replace('#field#', fieldConf['field']), fieldConf['search']);
         }
@@ -107,7 +111,7 @@ export class ServerDataSource extends LocalDataSource {
   }
 
   protected addPagerRequestOptions(requestOptions: RequestOptionsArgs): RequestOptionsArgs {
-    let searchParams: URLSearchParams = <URLSearchParams> requestOptions.search;
+    const searchParams: URLSearchParams = <URLSearchParams>requestOptions.params;
 
     if (this.pagingConf && this.pagingConf['page'] && this.pagingConf['perPage']) {
       searchParams.set(this.conf.pagerPageKey, this.pagingConf['page']);
