@@ -1,45 +1,4 @@
-function isSpecificValue(val) {
-  return (
-    val instanceof Buffer
-    || val instanceof Date
-    || val instanceof RegExp
-  ) ? true : false;
-}
-
-function cloneSpecificValue(val): any {
-  if (val instanceof Buffer) {
-    var x = new Buffer(val.length);
-    val.copy(x);
-    return x;
-  } else if (val instanceof Date) {
-    return new Date(val.getTime());
-  } else if (val instanceof RegExp) {
-    return new RegExp(val);
-  } else {
-    throw new Error('Unexpected situation');
-  }
-}
-
-/**
- * Recursive cloning array.
- */
-function deepCloneArray(arr): any {
-  var clone = [];
-  arr.forEach(function (item, index) {
-    if (typeof item === 'object' && item !== null) {
-      if (Array.isArray(item)) {
-        clone[index] = deepCloneArray(item);
-      } else if (isSpecificValue(item)) {
-        clone[index] = cloneSpecificValue(item);
-      } else {
-        clone[index] = deepExtend({}, item);
-      }
-    } else {
-      clone[index] = item;
-    }
-  });
-  return clone;
-}
+import { cloneDeep } from 'lodash';
 
 /**
  * Extending object that entered in first argument.
@@ -50,7 +9,7 @@ function deepCloneArray(arr): any {
  * object as first argument, like this:
  *   deepExtend({}, yourObj_1, [yourObj_N]);
  */
-export const deepExtend = function (...objects): any {
+export const deepExtend = function(...objects: Array<any>): any {
   if (arguments.length < 1 || typeof arguments[0] !== 'object') {
     return false;
   }
@@ -59,14 +18,14 @@ export const deepExtend = function (...objects): any {
     return arguments[0];
   }
 
-  var target = arguments[0];
+  const target = arguments[0];
 
   // convert arguments to array and cut off target object
-  var args = Array.prototype.slice.call(arguments, 1);
+  const args = Array.prototype.slice.call(arguments, 1);
 
-  var val, src, clone;
+  let val, src;
 
-  args.forEach(function (obj) {
+  args.forEach((obj: any) => {
     // skip argument if it is array or isn't object
     if (typeof obj !== 'object' || Array.isArray(obj)) {
       return;
@@ -90,12 +49,7 @@ export const deepExtend = function (...objects): any {
 
         // just clone arrays (and recursive clone objects inside)
       } else if (Array.isArray(val)) {
-        target[key] = deepCloneArray(val);
-        return;
-
-        // custom cloning and overwrite for specific objects
-      } else if (isSpecificValue(val)) {
-        target[key] = cloneSpecificValue(val);
+        target[key] = cloneDeep(val);
         return;
 
         // overwrite by new value if source isn't object or array
@@ -117,21 +71,21 @@ export const deepExtend = function (...objects): any {
 export class Deferred {
 
   promise: Promise<any>;
-  
+
   resolve: any;
   reject: any;
-  
-  constructor () {
+
+  constructor() {
     this.promise = new Promise((resolve, reject) => {
       this.resolve = resolve;
       this.reject = reject;
-    })
+    });
   }
 }
 
 // getDeepFromObject({result: {data: 1}}, 'result.data', 2); // returns 1
 export function getDeepFromObject(object = {}, name: string, defaultValue?: any) {
-  let keys = name.split('.');
+  const keys = name.split('.');
   // clone the object
   let level = deepExtend({}, object);
   keys.forEach((k) => {
